+++ composeApp/src/commonMain/composeResources/files/CHANGELOG.md
@@ -1,3 +1,11 @@
+Version: 1.3.2
+Date: 10-05-2026
+Changes:
+  - Install from remote ui is less of a mess
+Fixed:
+  - Adding repo would fail due to hidden characters
+  - Improper changelog divider
+----------------------------------------------------------------------------------------------------
 Version: 1.3.1
 Date: 09-05-2026
 Added:
@@ -11,7 +19,7 @@ Fixed:
 	- Application no longer checks for update every time it's opened when closed to tray
 	- Application update progress bar correctly reports progress
 	- Updates are now installed correctly (at least on windows)
----------------------------------------------------------------------------------------------------
+----------------------------------------------------------------------------------------------------
 Version: 1.3.0
 Date: 08-05-2026
 Added: